Try to fix tests
 On branch master
 Your branch is up to date with 'origin/master'.

 Changes to be committed:
	modified:   tests/test_spark_ai.py
	new file:   tests/test_spark_connect.py
